P9, P25, P26, P27, P28, P31, P32, P39, P40, P47, P48, P49, P50, P58, P59, P69, P70, P79: Simplified and tweaked Java solution for clarity.

+++ p028.java
@@ -26,10 +26,10 @@ public final class p028 implements EulerSolution {
 	private static final int SIZE = 1001;  // Must be odd
 	
 	public String run() {
-		int sum = 1;  // Special case for size 1
+		long sum = 1;  // Special case for size 1
 		for (int n = 3; n <= SIZE; n += 2)
 			sum += 4 * n * n - 6 * (n - 1);
-		return Integer.toString(sum);
+		return Long.toString(sum);
 	}
 	
 }
